requirements-lambda.txt
datetime
python-dateutil
requests
openpyxl
pymysql
boto
